feat: add mobile and desktop-specific custom emoji impls
