+++ shared/common-adapters/custom-emoji.d.ts
@@ -1,0 +1,12 @@
+import {Component} from 'react'
+
+type emojiSize = 'Small' | 'Medium' | 'Big'
+export type Props = {
+  size: emojiSize
+  src: string
+  alias?: string
+}
+
+declare const emojiSizes: {[key in emojiSize]: number}
+
+export default class Emoji extends Component<Props> {}
